Move rnd_chars to utils
